Move command switch to command.js
